chore: change webpack config for use the loader in demo files
